backend/requirements.txt
Flask>=2.0
Pillow>=9.0
numpy>=1.21
easyocr>=1.6
opencv-python-headless>=4.7
pdf2image>=1.16   # optional; required only if you want PDF uploads
requests>=2.28
